Add files via upload
Implemented a toggle for dark/light modes. Need to make a fix, though, as clicking to generate a new joke resets the color back to default. Began adding a way to choose between generating 1-4 joke/memes. Only the drop down bar has been added so far, as I need to research it more.